+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Updated [WET](https://github.com/wet-boew/wet-boew/releases/) to v4.0.47.1
 - Added an *acronym* parameter to the custom menu allowing users to provide a description for an abbreviation.  Please visit the sample pages for all the details.
 - Updated search by adding an accept-charset attribute
+- Updated ESDC, ECCC and Labour sub-themes to have a pre-configured layout
 - Internal changes and optimizations
 - Bug fixes